pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["app", "agents", "main", "pipeline", "tools"]

[project]
name = "re-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "html5lib>=1.1",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-openrouter>=0.2.1",
    "lxml>=5.0.0",
    "openai>=1.30.0",
    "orjson>=3.9.0",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.34.0",
    "tavily-python>=0.3.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.6.0",
]
